added powershell script and trimmed down bash script

+++ tasks.py
@@ -13,13 +13,14 @@ BUILD_DIR = os.path.join("./build")
 EXE="firmware.elf"
 FLASH_SCRIPT="flash.jlink"
 DEBUG_SCRIPT="debug.jlink"
+JLINK_ID=""
 
 @task
 def debug(c):
     """Debug with J-Link Edu Mini"""
     with c.cd("build"):
-        c.run("JLinkGDBServer -select USB=801031598 -device RP2040_M0_1 -endian little -if SWD -speed 4000 \
-                                -ir -noLocalhostOnly -nologtofile",pty=True)
+        c.run("JLinkGDBServer -select USB={} -device RP2040_M0_1 -endian little -if SWD -speed 4000 \
+                                -ir -noLocalhostOnly -nologtofile".format(JLINK_ID),pty=True)
         # c.run("arm-none-eabi-gdb {}".format(EXE),pty=True)
         # c.run('kill $(pgrep JLinkGDBServer)',pty=True)
 
@@ -27,7 +28,10 @@ def debug(c):
 @task
 def clean(c):
     """Clean all binaries build from project"""
-    c.run("rm -rf build/*".format(EXE)) 
+    if os.name == 'nt':
+        c.run("DEL /F /A /Q  build")
+    else:
+        c.run("rm -rf build/*".format(EXE)) 
     
 @task
 def test(c):
@@ -43,7 +47,10 @@ def cmake(c):
 def build(c):
     """Build the project"""
     with c.cd(BUILD_DIR):
-        c.run("make -j4")
+        if os.name == 'nt':
+            c.run("ninja -j4")
+        else:
+            c.run("make -j4")
 
 @task(pre=[build])
 def flash(c):
